renamed commands.js to getAPIs.js for clarity
